inserido popupMenu ao card de categoria
